Avoid duplicate error report when renaming HsDoc stuff
This patch is a bit of a hack to avoid a duplicate error when checking
	class C a where
	  op :: a -> a
	  op2 x = x
(This is tcfail077)

For reasons I don't understand, the decl of op2 generates an HsDeclEntity,
and that gives a renamer error which duplicates the (better) one that 
comes from rnMethodBinds.

A better fix might be to get rid of HsDeclEntities altogether.