Log-message for: Improve command-line parser (add OptIntSuffix); make -fliberate-case-threshold dynamic
(Alas, Darcs failed to record my log-message for the above patch, 
so this patch is an attempt to add the log message retrospectively.)

Parsing options numeric arguments, such as
	-fliberate-case-threshold=30
for dynamic flags wasn't being handled well.  I elaborated the
command-line parser to have a new constructor, OptIntSuffix, for this
case, and did the consequential changes.

This patch also makes the -fliberate-case-threshold flag into a
dynamic flag.