Improve simplification of coercions
At the moment GHC really does very little simplification of coercions.
This patch improves matters, but it's still not great, especially when
you have chains linked together with 'trans'. 

I'm also concerned that I have not yet implemented the 'leftc' and 'rightc'
coercions we added to the paper. 

But at least things are better than they were.  In particular
	g `trans` sym g
now cancels to give the identity.