fix adjustor generation on ia64 (test case ffi009)
Some fixes to adjustor functions.  The 8-byte address returned by the 
allocator is adjusted to be aligned to 16-byte boundaries.  Fixed a typo 
in inserting an immediate address into an instruction.

This fixes the calls to 5-argument and 6-argument functions in ffi009.  
Some functions still break.  I suspect it's related to passing arguments on 
the stack.