Make ticky work, at least partly, on 64-bit machines
The ticky StgEntCounter structure was trying to be clever by using a
fixed-width 32-bit field for the registeredp value.  But the code generators
are not up to handling structures packed tightly like this (on a 64-bit
architecture); result seg-fault on 64-bit.

Really there should be some complaint from the code generators, not simply
a seg fault.

Anyway I switched to using native words for StgEntCounter fields, and
now at least it works.