Dont tidy up tyvars after :print type reconstruction
I introduced a bug yesterday when I changed the way tidying up was performed.
As a result of tidying,  cvObtainTerm could be returning types 
with regular tyvars inside, which never should.

But actually, it's better if we do not do the tidying up, in order to 
keep the tyvar names from the environment.
New names will be introduced only when an existential is found, which
is not so common. In this case the user will see a funny name. 
Is that really an issue?