Refactoring, tidyup and improve layering
The stack of breakpoint resume contexts is now part of the
InteractiveContext and managed by the GHC API.  This prevents misuse
of the resume context by the client (e.g. resuming a breakpoint that
isn't the topmost, which would lead to a confused IC at the least).

I changed the TypeEnv in the IC to a [Id].  It only contained Ids
anyway, and this allows us to have shadowing, which removes an ugly
and annoying restriction.

The parts of the GHC API which deal with interactive evaluation are
now in a module of their own, InteractiveEval.