Make -frewrite-rules into a dynamic flag; off for -O0
Argubly rewrite rules should not fire with -O0, and it turns
out that when compiling GHC.Base with -O0 we get a crash if
the rewrite rules do fire (see Note [Scoping for Builtin rules]
in PrelRules).

So unless someone yells, rewrite rules are off with -O0.

The new (now dynamic) flag is 
    -frewrite rules (with -fno-rewrite-rules to disable)

The old (static) flag -frules-off is gone.