Improve full laziness by floating allocations out of value lambdas
This patch simplifies the code slightly, and simultaneously improves
full laziness by floating allocations (lambdas, constructor apps) out
of loops.

See Note [Escaping a value lambda] in SetLevels, which explains.

There's a test that shows it up: simplrun009

This relevant to SpecConstr, because a call looks like
	f lvl
instead of
	f (\x. blah)
and the latter is easier to match in a robust way.