FIX read040: patterns with type sig on LHS of do-binding
f () = do { x :: Bool <- return True; ... }

For some reason the production for 'pat' required 'infixexp' on the
LHS of a do-notation binding.  This patch makes it an 'exp', which 
thereby allows an expression with a type sig.

Happily, there are no new shift-reduce errors, so I don't think this
will break anything else.