Print infix type constructors in an infix way
Fixes Trac #1425.  The printer for types doesn't know about fixities.
(It could be educated to know, but it doesn't at the moment.)  So it
treats all infix tycons as of precedence less than application and function
arrrow.

I took a slight shortcut and reused function-arrow prededence, so I think
you may get
	T -> T :% T
meaning
	T -> (T :% T)

If that becomes a problem we can fix it.